refactoring application files
