+++ clients/core/src/App.tsx
@@ -11,7 +11,7 @@ import ImprintPage from './LegalPages/Imprint'
 import AboutPage from './LegalPages/AboutPage'
 import { CourseOverview } from './Course/CourseOverview'
 import { ApplicationLoginPage } from './Application/ApplicationLoginPage'
-import { ApplicationAuthenticated } from './Application/ApplicationAuthenticated'
+import { ApplicationAuthenticated } from './Application/pages/ApplicationAuthenticated/ApplicationAuthenticated'
 import { Toaster } from '@/components/ui/toaster'
 import CourseConfiguratorPage from './CourseConfigurator/CourseConfiguratorPage'
 import { PermissionRestriction } from './management/PermissionRestriction'
