+++ clients/core/src/Application/ApplicationLoginPage.tsx
@@ -8,7 +8,7 @@ import { LoadingState } from './components/LoadingState'
 import { ErrorState } from './components/ErrorState'
 import { useState } from 'react'
 import { ApplicationLoginCard } from './components/ApplicationLoginCard'
-import { ApplicationFormView } from './ApplicationFormView'
+import { ApplicationFormView } from './pages/ApplicationForm/ApplicationFormView'
 import { Student } from '@/interfaces/student'
 import { CreateApplicationAnswerText } from '@/interfaces/application_answer_text'
 import { CreateApplicationAnswerMultiSelect } from '@/interfaces/application_answer_multi_select'
