+++ clients/core/src/Application/pages/ApplicationAuthenticated/ApplicationAuthenticated.tsx
@@ -1,20 +1,20 @@
 import { useParams } from 'react-router-dom'
-import { AuthenticatedPageWrapper } from '../components/AuthenticatedPageWrapper'
+import { AuthenticatedPageWrapper } from '../../../components/AuthenticatedPageWrapper'
 import { ApplicationFormWithDetails } from '@/interfaces/application_form_with_details'
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query'
-import { getApplicationFormWithDetails } from '../network/queries/applicationFormWithDetails'
-import { LoadingState } from './components/LoadingState'
-import { ErrorState } from './components/ErrorState'
-import { ApplicationHeader } from './components/ApplicationHeader'
-import { ApplicationFormView } from './ApplicationFormView'
+import { getApplicationFormWithDetails } from '../../../network/queries/applicationFormWithDetails'
+import { LoadingState } from '../../components/LoadingState'
+import { ErrorState } from '../../components/ErrorState'
+import { ApplicationHeader } from '../../components/ApplicationHeader'
+import { ApplicationFormView } from '../ApplicationForm/ApplicationFormView'
 import { useAuthStore } from '@/zustand/useAuthStore'
 import { Student } from '@/interfaces/student'
-import { getApplication } from '../network/queries/application'
+import { getApplication } from '../../../network/queries/application'
 import { GetApplication } from '@/interfaces/get_application'
 import { PostApplication } from '@/interfaces/post_application'
-import { postNewApplicationAuthenticated } from '../network/mutations/postApplicationAuthenticated'
+import { postNewApplicationAuthenticated } from '../../../network/mutations/postApplicationAuthenticated'
 import { useState } from 'react'
-import { ApplicationSavingDialog } from './components/ApplicationSavingDialog'
+import { ApplicationSavingDialog } from '../../components/ApplicationSavingDialog'
 import { CreateApplicationAnswerText } from '@/interfaces/application_answer_text'
 import { CreateApplicationAnswerMultiSelect } from '@/interfaces/application_answer_multi_select'
 import { InfoBanner } from './components/InfoBanner'
