+++ clients/core/src/Application/pages/ApplicationAuthenticated/components/InfoBanner.tsx
