+++ clients/core/src/Application/pages/ApplicationForm/components/MultiSelectForm/ApplicationQuestionMultiSelectForm.tsx
@@ -2,12 +2,12 @@ import React, { forwardRef, useImperativeHandle } from 'react'
 import { useForm } from 'react-hook-form'
 import { zodResolver } from '@hookform/resolvers/zod'
 import { Form, FormControl, FormField, FormItem, FormMessage } from '@/components/ui/form'
-import { QuestionMultiSelectFormRef } from '../utils/QuestionMultiSelectFormRef'
+import { QuestionMultiSelectFormRef } from '../../utils/QuestionMultiSelectFormRef'
 import { ApplicationQuestionMultiSelect } from '@/interfaces/application_question_multi_select'
 import { createValidationSchema } from './validationSchema'
 import { CheckboxQuestion } from './CheckboxQuestion'
 import { MultiSelectQuestion } from './MultiSelectQuestion'
-import { checkCheckBoxQuestion } from '../utils/CheckBoxRequirements'
+import { checkCheckBoxQuestion } from '../../utils/CheckBoxRequirements'
 
 interface ApplicationQuestionMultiSelectFormProps {
   question: ApplicationQuestionMultiSelect
