+++ clients/core/src/Application/pages/ApplicationForm/components/MultiSelectForm/validationSchema.ts
