+++ clients/core/src/Application/pages/ApplicationForm/utils/CheckBoxRequirements.ts
