+++ clients/core/src/ApplicationAdministration/pages/ApplicationAssessment/components/ApplicationManualAddingDialog/ApplicationManualAddingDialog.tsx
@@ -16,7 +16,7 @@ import { Alert, AlertDescription, AlertTitle } from '@/components/ui/alert'
 import { ApplicationForm } from '@/interfaces/application_form'
 import { getApplicationForm } from '../../../../../network/queries/applicationForm'
 import { UniversitySelection } from './components/UniversitySelection'
-import { ApplicationFormView } from '../../../../../Application/ApplicationFormView'
+import { ApplicationFormView } from '../../../../../Application/pages/ApplicationForm/ApplicationFormView'
 import { StudentSearch } from './components/StudentSearch'
 import { Student } from '@/interfaces/student'
 import { postNewApplicationManual } from '../../../../../network/mutations/postApplicationManual'
