+++ clients/core/src/ApplicationAdministration/pages/ApplicationConfiguration/ApplicationQuestionConfig/FormPages/ApplicationQuestionCard.tsx
@@ -28,7 +28,7 @@ import { Button } from '@/components/ui/button'
 import { DeleteConfirmation } from '../components/DeleteConfirmation'
 import { questionsEqual } from '../handlers/computeQuestionsModified'
 import { QuestionStatus, QuestionStatusBadge } from '../components/QuestionStatusBadge'
-import { checkCheckBoxQuestion } from '../../../../../Application/utils/CheckBoxRequirements'
+import { checkCheckBoxQuestion } from '../../../../../Application/pages/ApplicationForm/utils/CheckBoxRequirements'
 import { DescriptionMinimalTiptapEditor } from '@/components/minimal-tiptap/form-description-tiptap'
 import { TooltipProvider } from '@/components/ui/tooltip'
 import { Switch } from '@/components/ui/switch'
